commit41

+++ demo/bill/views.py
@@ -490,7 +490,7 @@ def add_invoice(request):
         for i in range(1,no_of_items+1):
             item_details_id=request.POST.get('item'+str(i))
             item_details=item.objects.get(id=item_details_id)
-            quantity=int(request.POST.get('quantity'+str(i)))
+            quantity=float(request.POST.get('quantity'+str(i)))
             rate=float(request.POST.get('rate'+str(i)))
             amount=quantity*rate
             taxable_amt+=amount
